Is the body warmer patch useful for dysmenorrhea?
It’s useful. Every time menstruation comes, touch your belly with your hand and feel that the whole belly is cold. If you put a hot water bottle or body warmer, the whole belly will feel much warmer, and people will not be so uncomfortable. It is also important to take more rest and maintain a happy mood. Pay attention to hygiene when you are menstruating. It is best not to touch cold water during this period, and do not eat spicy or cold things. I used to soak brown sugar water when I had dysmenorrhea. If I wanted to do it, I would cook brown sugar ginger water. I felt that it would be effective after drinking it. I used to drink it every month, but now it’s much better. I only drink it occasionally.